acciones actualizado con el no reload
